+++ event/src/main/java/com/pyokemon/event/EventApplication.java
@@ -5,9 +5,7 @@ import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.scheduling.annotation.EnableScheduling;
 
-@SpringBootApplication(scanBasePackages = {"com.pyokemon"},
-    exclude = {
-        org.springframework.boot.autoconfigure.security.servlet.SecurityAutoConfiguration.class})
+@SpringBootApplication(scanBasePackages = {"com.pyokemon"})
 @MapperScan("com.pyokemon.event.repository")
 @EnableScheduling
 public class EventApplication {
