+++ event/src/main/java/com/pyokemon/event/dto/EventDetailResponseDTO.java
@@ -27,6 +27,5 @@ public class EventDetailResponseDTO {
 
   // venue
   private String venueName;
-
   private boolean isSaved;
 }
