+++ event/src/main/java/com/pyokemon/event/dto/EventRegisterDto.java
@@ -27,6 +27,7 @@ public class EventRegisterDto {
   private Long eventId;
 
   // @NotNull(message = "Account ID is required") // 임시로 주석 처리 (account 서비스 완료 전까지)
+  @NotNull(message = "Account ID is required")
   private Long accountId;
 
   @NotBlank(message = "Title is required")
