+++ event/src/main/java/com/pyokemon/event/dto/MonthlyEventSummaryResponse.java
@@ -1,0 +1,19 @@
+package com.pyokemon.event.dto;
+
+import java.util.List;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class MonthlyEventSummaryResponse {
+    private List<MonthlyEventDTO> events;
+    private MonthlySummaryDTO summary;
+}
