+++ event/src/main/java/com/pyokemon/event/dto/MonthlySummaryDTO.java
@@ -1,0 +1,18 @@
+package com.pyokemon.event.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class MonthlySummaryDTO {
+  private int totalRevenue;
+  private int activeEventCount;
+  private int totalTicketsSold;
+}
