+++ event/src/main/resources/application.yml
@@ -27,10 +27,3 @@ spring:
 mybatis:
   mapper-locations: classpath:mapper/*.xml
 
----
-spring:
-  config:
-    activate:
-      on-profile: "dev"
-    import:
-      - classpath:dev-config.yml
